+++ CHANGELOG.rst
@@ -3,6 +3,8 @@ Changes since 0.6.0 onwards.
 
 0.9.11 (2021-04-24)
 ~~~~~~~~~~~~~~~~~~~
+* Add result.lnlike to output of dataset and multivisit fitting routines.
+* Add log_sigma keyword to dataset.lmfit_transit and dataset.lmfit_eclipse
 * Correct calculation of result.chisqr in dataset.lmfit_transit and
   dataset.lmfit_eclipse
 * Catch invalid combinations of b and k in dataset._log_prior() and remove